pyproject.toml
[project]
name = "umashow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "icnsutil>=1.1.0",
    "pillow>=12.0.0",
]
